requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-telegram-bot
openai==1.3.7
sqlalchemy==2.0.23
pydantic==2.5.1
pydantic-settings==2.1.0
aiohttp==3.9.1
python-multipart==0.0.6
python-dotenv==1.0.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
psycopg2-binary==2.9.10
groq
